Update settings.py
Environment variables for this particular database are in no way needed, because it has no actual information.
Database only has information about programming languages (name, paradigm and created by) and faangm (name, founders, created and location).
Like I said before using environment variables for this db is not needed, but I still chose to add env var's because it is very bad practice to directly add password
to file (when working with actual user data) so I wanted to include it here. I don't want people to use this method and also to show that I know and do use environment variables.